move get_exports_by_form to analytics
rather than dbaccessors since it uses stale ok
